refactor: reorganize product batches folders by layer (handler, repo, service)
