+++ internal/router/sections.go
@@ -2,11 +2,11 @@ package router
 
 import (
 	"github.com/go-chi/chi/v5"
-	"github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler"
+	productBatchHandler "github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/product_batch"
 	sectionHandler "github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/section"
 )
 
-func MountSectionRoutes(api chi.Router, hd *sectionHandler.SectionDefault, hdPB *handler.ProductBatchesHandler) {
+func MountSectionRoutes(api chi.Router, hd *sectionHandler.SectionDefault, hdPB *productBatchHandler.ProductBatchesHandler) {
 	api.Route("/sections", func(r chi.Router) {
 		r.Get("/", hd.FindAllSections)
 		r.Post("/", hd.CreateSection)
